refactor factory클래스도 도메인으로 이동.
