+++ src/main/java/racingcar/RacingGameLauncher.java
@@ -2,6 +2,7 @@ package racingcar;
 
 import java.util.List;
 import racingcar.domain.RacingGame;
+import racingcar.domain.RacingGameFactory;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
