+++ src/main/java/racingcar/domain/RacingGameFactory.java
@@ -1,13 +1,7 @@
-package racingcar;
+package racingcar.domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import racingcar.domain.Car;
-import racingcar.domain.CarMovingCondition;
-import racingcar.domain.RacingCars;
-import racingcar.domain.RacingGame;
-import racingcar.domain.RandomCarMovingCondition;
-import racingcar.domain.TimeBaseRandomGenerator;
 
 public class RacingGameFactory {
