+++ src/test/java/racingcar/domain/RacingGameTest.java
@@ -9,10 +9,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.RacingGameFactory;
-import racingcar.domain.Car;
-import racingcar.domain.RacingCars;
-import racingcar.domain.RacingGame;
 
 public class RacingGameTest {
